+++ application/static/pics/atlas_title.png
