+++ application/templates/Lex_Ejoris.html
@@ -11,10 +11,13 @@
 
 </head>
 <body>
-<body bgcolor="#00000">
 
 <a href="{{ url_for('Lex_Ejoris_Tree') }}" target="_blank">
 <img align="right" width="640" height"360" src ="{{ LE_image }}"></a>
+
+
+
+
 
 
 </body>
