+++ application/templates/home.html
@@ -5,20 +5,18 @@
 <center><title>Atlas Companion</title><center>
 <center><img src= "{{title_image}}"></center>
 <br>
-<br>
 <center><h3 style="color:#fa0015";>Click to Navigate</h3><center>
 
-<h1 style="color:#ffffff"; align="left"> Divination Cards</h1>
 
 </head>
 <body>
 <body bgcolor="#00000">
 <a href="{{ url_for('Divination_Cards') }}">
-<img align="left" width="426" height="240" src= "{{div_card_icon}}" vspace="20"></a>
+<img align="left" width="250" height="125" hspace="160" src= "{{div_card_icon}}"></a>
 
 
 
-<center><img src="{{home_image}}" usemap="#image-map" vspace="100"></center>
+<center><img src="{{home_image}}" usemap="#image-map" vspace="50"></center>
 
 <map name="image-map">
     <area target="" alt="Tirn's End" title="Tirn's End" href="{{ url_for('Tirns_End') }}" coords="434,237,573,155,635,134,659,189,676,245,665,319,735,381,658,432,627,507,520,447,473,449,428,465,352,464,284,449,240,398,229,345,338,290,390,257" shape="poly">
