+++ create.py
@@ -1,0 +1,3 @@
+from application import db
+
+db.create_all()
