--- templates/New_Vastir.html
@@ -1,21 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-
-<title>New Vastir</title>
-
-<center><img src ="{{ NV_Title }}"><center>
-
-<a href="{{ url_for('home') }}">
-<img align="left" width "110" height="90" src="{{ HomeB }}"></a>
-
-</head>
-<body>
-
-<a href="{{ url_for('New_Vastir_Tree') }}" target="_blank">
-<img align="right" width="640" height="360" src ="{{ NV_image }}"></a>
-
-	 
-</body>
-</html>
-
